requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
APScheduler==3.6.3
async-timeout==4.0.2
attrs==22.2.0
Babel==2.9.1
cachetools==4.2.2
certifi==2022.12.7
charset-normalizer==3.0.1
class-registry==2.1.2
databases==0.7.0
dispatcher==1.0
filters==1.3.2
frozenlist==1.3.3
greenlet==2.0.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.16.3
httpx==0.23.3
hyperframe==6.0.1
idna==3.4
magic-filter==1.0.9
multidict==6.0.4
orm==0.3.1
psycopg2-binary==2.9.5
pyTelegramBotAPI==4.10.0
python-dateutil==2.8.2
python-dotenv==0.21.1
python-telegram-bot==13.7
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
regex==2022.10.31
requests==2.28.2
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.46
telebot==0.0.5
tornado==6.2
typesystem==0.3.1
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
yarl==1.8.2
